🐛 fix: share interest normalization helpers
